% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitISUVB.R
\name{fitISUVB}
\alias{fitISUVB}
\title{Generic function for IS-UVB gradient ascent}
\usage{
fitISUVB(lambda, qDist, samples, dSamples, logjoint, maxIter = 5000,
  alpha = 0.01, beta1 = 0.9, beta2 = 0.99, threshold = 0.01, ...)
}
\arguments{
\item{lambda}{The vector or one column matrix to be optimised}

\item{qDist}{A function with first two arguments: samples, lambda, that returns a list with two elements: grad, the gradient of log Q, and val, the value of log Q.}

\item{samples}{The matrix of samples from the previous UVB distribution. Each row must be a single draw from a multivariate distribution}

\item{dSamples}{A vector of the density of each row of samples according the the previous UVB distribution}

\item{logjoint}{A vector of the log of the joint density between the data and each row of samples}

\item{maxIter}{Interger. Maximum number of gradient ascent iterations.}

\item{alpha}{adam optimisation control parameter.}

\item{beta1}{adam optimisation control parameter.}

\item{beta2}{adam optimisation control parameter.}

\item{threshold}{Maximum difference in mean value of ELBO before convergence is achieved.}

\item{...}{Extra arguments passed to qDist}

\item{rollingWindowSize}{Integer. Take the mean of this many most recent iterations to assess convergence. Defaults to 5}
}
\description{
Returns the vector of optimal lambda values, the history of ELBO values, the number of iterations to converge, and the history of the effective sample size
}
