% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianDensity.R
\name{gaussianDensity}
\alias{gaussianDensity}
\title{Calculates the marginal density of a transformed multivariate gaussian.}
\usage{
gaussianDensity(mu, sigma, transform = NULL, names = NULL,
  supports = NULL)
}
\arguments{
\item{mu}{the length n mean vector of the multivariate gaussian}

\item{sigma}{the length n vector of standard deviations.}

\item{transform}{a length n vector with the marginal transform of each variable from c('identity', 'exp', 'sigmoid', stretchedSigmoid'). If null, defaults each variable to 'identity'}

\item{names}{a length n vector of the names of each variable. If null, defaults to V1, V2, .., Vn}

\item{supports}{a matrix with n columns that gives the support of each variable (per column) to calculate the densities over. If NULL, automatically find a suitable density.}
}
\description{
This function returns a dataframe with columns for the support of each marginal distribution, the density at that value, and the name of that variable if provided.
}
