% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianVB.R
\name{gaussianVB}
\alias{gaussianVB}
\title{Reparameterised VB for a (Mixture) Gaussian distribution}
\usage{
gaussianVB(data, lambda, model, dimEpsilon, S = 25, zEpsilon = TRUE,
  RQMC = TRUE, batch = 0, maxIter = 5000, alpha = 0.01,
  beta1 = 0.9, beta2 = 0.99, rollingWindowSize = 5,
  threshold = 0.01, ...)
}
\arguments{
\item{data}{Data passed directly into model.}

\item{lambda}{Parameter vector to be optimised. If lambda is a matrix, each column is assumed to be the parameter vector for a component of a mixture of multivariate gaussians, with the first row corresponding to unnormalised weights z, where the normalised weights are given by p = exp(z) = sum(exp(z))}

\item{model}{A function with first three arguments: data, lambda, epsilon that returns a list with elements 'grad', the vector of the gradient of the ELBO with respect to lambda, and 'val', the value of the ELBO for this lambda.}

\item{dimEpsilon}{Integer. The dimension of the auxilliary random variable epsilon per mixture component}

\item{S}{Integer. The number of Monte Carlo estimates per iteration. Defaults to 25.}

\item{zEpsilon}{Boolean, if true simualte epsilon from a standard normal distribution. If false simulate epsilon from a uniform distribution Defaults to TRUE.}

\item{RQMC}{Boolean, if true simulate epsilon from Randomised Quasi Monte Carlo based off the sobol seqence. Defaults to TRUE.}

\item{batch}{Integer. If data is a matrix, calculate gradient based on batch many columns per iteration, if ncol(data) is divisible by batch. If batch = 0 use the whole data object per iteration.}

\item{maxIter}{Interger. Maximum number of gradient ascent iterations. Defaults to 5000.}

\item{alpha}{adam optimisation control parameter. Defaults to 0.01.}

\item{beta1}{adam optimisation control parameter. Defaults to 0.9.}

\item{beta2}{adam optimisation control parameter. Defaults to 0.99.}

\item{rollingWindowSize}{Integer. Take the mean of this many most recent iterations to assess convergence. Defaults to 5 if batch = 0, or ncol(data) / batch otherwise.}

\item{threshold}{Maximum difference in mean value of ELBO before convergence is achieved. Defaults to 0.01.}

\item{...}{Extra arguments passed into model}
}
\description{
This function runs a Variational Bayes gradient ascent to find the parameters of a mixture of multivariate gaussians distrbituion
q that minimises the KL divergence to the true posterior.
Given a model function that calculates the derivative of the ELBO with resepct to lambda, it will return a list containing the optimal lambda,
history of ELBO values, and iterations required to converge for a given model function and dataset.
}
