% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolisHastings.R
\name{metropolisHastings}
\alias{metropolisHastings}
\title{Block Random Walk Metrolpolis Hastings Sampler}
\usage{
metropolisHastings(start, iter, model, targetAcceptance = 0.234,
  burnIn = 0, thin = 1, stepsize = 0.01, ...)
}
\arguments{
\item{start}{A vector of starting values for the parameters. If a named vector is inputted, the names will be attached to the output}

\item{iter}{The total number of iterations,}

\item{model}{A function with first argument theta that evaluates the log joint density at that value of theta}

\item{targetAcceptance}{The targetted MH acceptance rate, defaults to 0.234}

\item{burnIn}{The number of discarded iterations, should be lower than iter. Defaults to 0.}

\item{thin}{The thinning factor. Draws will only be saved when the iteration number equals zero, modulus thin. Defaults to 1.}

\item{stepsize}{The starting standard deviation for the Random Walk sampler. Will change dynamically through the algorithm}

\item{...}{Extra arguments passed to model, eg. Data, Prior parameters}
}
\description{
Jointly samples a candidate from a Random Walk Metrolpolis Hastings algorithm with a given acceptance rate to be targeted.
}
