% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vbetelMatrixCalculations}
\alias{vbetelMatrixCalculations}
\title{Returns the derivative of the log-joint distribution with respect to theta for VBETEL}
\usage{
vbetelMatrixCalculations(g, hessian, lambdaHat, exponent, dgdt)
}
\arguments{
\item{g}{An N x P matrix where rows correspond to the P dimensional moment condition evaluated at each of the N observations.}

\item{hessian}{The P x P matrix of the hessian matrix of h(theta, lambda), usually provided by nlm(...)$hessian.}

\item{lambdaHat}{A P dimensional vector of the optimal lambda values.}

\item{exponent}{An N dimensional vector of exp(lambdaHat * g) evaluated at each of the N observations}

\item{gdt}{A P x P x N array of the P x P matrix of the derivative of g with respect to theta, each slice corresponds to one array}
}
\description{
Given the intermediate inputs, this function calculates the last few stages of the derivative of the logjoint with respect to theta.
}
