% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vbetelMatrixCalculations}
\alias{vbetelMatrixCalculations}
\title{Returns the derivative of the log-joint distribution with respect to theta for VBETEL}
\usage{
vbetelMatrixCalculations(g, hessian, lambdaHat, exponent, dgdt)
}
\arguments{
\item{g}{An N x D matrix where rows correspond to the D dimensional moment condition evaluated at each of the N observations.}

\item{hessian}{The D x D matrix of the hessian matrix of h(theta, lambda), usually provided by nlm(...)$hessian.}

\item{lambdaHat}{A D dimensional vector of the optimal lambda values.}

\item{exponent}{An N dimensional vector of exp(lambdaHat * g) evaluated at each of the N observations}

\item{gdt}{A D x P x N array of the D x P matrix of the derivative of g (D dimensional) with respect to theta (P dimensional), 
each slice corresponds to the derivative matrix for one observation}
}
\description{
Given the intermediate inputs, this function calculates the last few stages of the derivative of the logjoint with respect to theta.
}
